+++ ui-admin-web/src/components/reserved/ReservedBooksContent.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+export default function ReservedBooksContent() {
+  return (
+    <div>
+        
+    </div>
+  )
+}
