+++ ui-admin-web/src/config/messages/fr.json
@@ -148,16 +148,16 @@
     "content": "Vos livres seront prêts pour vous dans votre chambre. Merci d'utiliser notre service",
     "goBackBtn": "Retour à l'accueil"
   },
-  "returnBook": {
-    "yourRentedBooks": "Vos livres loués",
-    "bookToReturn": "Quel livre voulez-vous retourner aujourd'hui ?",
-    "book": "livres",
-    "rentAnotherBook": "Louer un autre livre",
-    "back": "retour",
-    "return": "retourner",
-    "returnDate": "date de retour",
-    "noRentedBooks": "Vous n'avez pas encore de livre loué. Découvrez la meilleure option dans votre bibliothèque personnelle maintenant.",
-    "rentBook": "Louer un livre"
+  "reservedBooks": {
+    "moveToBuy": "Déplacer pour acheter",
+    "moveToReadInHotel": "Déplacer pour lire à l'hôtel",
+    "readInHotel": "Lire à l'hôtel",
+    "returnDate": "Date de retour",
+    "reservedBooks1": "Vos",
+    "reservedBooks2": "livres",
+    "reservedBooks3": "réservés",
+    "modifyReservation": "Modifiez votre réservation comme vous le souhaitez",
+    "reserveAnotherBook": "Réserver un autre livre"
   },
   "depositPage": {
     "depositBookTitle1": "Déposez ce livre sur",
